source: extract inline javascript to a separate file.
This is mainly a cosmetic change, but it makes actually editing the
files a lot easier, in my opinion.

I also added the `defer` attribute to all of the script tags. This means
that the HTML renderer doesn't block while they're being downloaded,
which in turn theoretically means that the page load will appear to be
slightly more responsive.